reorg codegen
